+++ changelog.md
@@ -1,2 +1,3 @@
 - Fixed archeology loot modifiers breaking
 - Fixed the communion music disc interaction not working with Etched's album jukebox
+- Added config options for vanilla recipe changes